requirements.txt
openai
deepgram-sdk==3.8.0
python-dotenv
httpx
tqdm